requirements.txt
fastapi[standard]
Pillow
pymupdf
rembg[cpu]
pdf2docx